# cuda ggttgg performance fixes and studies