--- a/.gitignore
+++ b/.gitignore
@@ -7,6 +7,7 @@ build.*
 .build.*
 
 #=== Profiling ===
+nsight_logs
 profile_*
 *.qdrep
 *.sqlite