--- a/epochX/cudacpp/.gitignore
+++ b/epochX/cudacpp/.gitignore
@@ -1,2 +1,0 @@
-build.*
-.build.*