--- a/epochX/cudacpp/CODEGEN/PLUGIN/CUDACPP_SA_OUTPUT/madgraph/iolibs/template_files/gpu/process_matrix.inc
+++ b/epochX/cudacpp/CODEGEN/PLUGIN/CUDACPP_SA_OUTPUT/madgraph/iolibs/template_files/gpu/process_matrix.inc
@@ -2,7 +2,7 @@
       // *** COLOR ALGEBRA BELOW ***
       // (This method used to be called %(process_class_name)s::matrix_%(proc_name)s(%(matrix_args)s)?)
 
-      // The color matrix
+      // The color matrix [NB do keep 'static' for these constexpr arrays, see issue #283]
 %(color_matrix_lines)s
 
       // Sum and square the color flows to get the matrix element