--- a/epochX/cudacpp/CODEGEN/diffCode.sh
+++ b/epochX/cudacpp/CODEGEN/diffCode.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-diff --no-dereference -x '*log.txt' -x '*.o' -x '*.o.*' -x '*.a' -x '*.exe' -x 'lib' -x 'build.*' -x '.build.*' -x '*~' $* 
+diff --no-dereference -x '*log.txt' -x 'nsight_logs' -x '*.o' -x '*.o.*' -x '*.a' -x '*.exe' -x 'lib' -x 'build.*' -x '.build.*' -x '*~' $*