--- a/epochX/cudacpp/ee_mumu.auto/CODEGEN_ee_mumu_log.txt
+++ b/epochX/cudacpp/ee_mumu.auto/CODEGEN_ee_mumu_log.txt
@@ -116,6 +116,6 @@ INFO: /data/avalassi/GPU2020/MG5aMC/2.7.0_gpu/CODEGEN_ee_mumu/src/. and /data/av
 [1;32mDEBUG:  Entering PLUGIN_ProcessExporter.finalize [1;30m[output.py at line 154][0m [0m
 quit
 
-real	0m1.029s
-user	0m0.882s
-sys	0m0.120s
+real	0m1.333s
+user	0m0.884s
+sys	0m0.110s