… epochX/cudacpp/CODEGEN/generateAndCompare.sh
… epochX/cudacpp/ee_mumu.auto/SubProcesses/Makefile
… epochX/cudacpp/ee_mumu.auto/SubProcesses/P1_Sigma_sm_epem_mupmum/CPPProcess.cc
… epochX/cudacpp/ee_mumu.auto/SubProcesses/profile.sh
… epochX/cudacpp/ee_mumu.auto/SubProcesses/runTest.cc
… epochX/cudacpp/ee_mumu.auto/src/HelAmps_sm.cc
… epochX/cudacpp/ee_mumu.auto/src/Makefile
… epochX/cudacpp/ee_mumu/SubProcesses/Makefile
… epochX/cudacpp/ee_mumu/SubProcesses/P1_Sigma_sm_epem_mupmum/CPPProcess.cc
… epochX/cudacpp/ee_mumu/SubProcesses/profile.sh
… epochX/cudacpp/ee_mumu/SubProcesses/runTest.cc
… epochX/cudacpp/ee_mumu/src/HelAmps_sm.cc
… epochX/cudacpp/ee_mumu/src/Makefile
… epochX/cudacpp/gg_tt.auto/CODEGEN_gg_tt_log.txt
… epochX/cudacpp/gg_tt.auto/SubProcesses/Makefile
… epochX/cudacpp/gg_tt.auto/SubProcesses/P1_Sigma_sm_gg_ttx/CPPProcess.cc
… epochX/cudacpp/gg_tt.auto/SubProcesses/profile.sh
… epochX/cudacpp/gg_tt.auto/SubProcesses/runTest.cc
… epochX/cudacpp/gg_tt.auto/src/HelAmps_sm.cc
… epochX/cudacpp/gg_tt.auto/src/Makefile
… epochX/cudacpp/gg_ttgg.auto/CODEGEN_gg_ttgg_log.txt
… epochX/cudacpp/gg_ttgg.auto/SubProcesses/Makefile
… epochX/cudacpp/gg_ttgg.auto/SubProcesses/P1_Sigma_sm_gg_ttxgg/CPPProcess.cc
… epochX/cudacpp/gg_ttgg.auto/SubProcesses/profile.sh
… epochX/cudacpp/gg_ttgg.auto/SubProcesses/runTest.cc
… epochX/cudacpp/gg_ttgg.auto/src/HelAmps_sm.cc
… epochX/cudacpp/gg_ttgg.auto/src/Makefile
… epochX/cudacpp/gg_ttgg/SubProcesses/Makefile
… epochX/cudacpp/gg_ttgg/SubProcesses/profile.sh
… epochX/cudacpp/gg_ttgg/SubProcesses/runTest.cc