This is a WIP PR for cuda ggttgg performance studies (issue #277)

```
    Using the latest code, note that cuda110 (requires gcc92) was ~30% faster than cuda114
    
    On itscrd70.cern.ch [CPU: Intel(R) Xeon(R) Silver 4216 CPU] [GPU: 1x Tesla V100S-PCIE-32GB]:
    =========================================================================
    runExe /afs/cern.ch/user/a/avalassi/GPU2020/madgraph4gpuX/epochX/cudacpp/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg/build.none_d_inl0/gcheck.exe -p 64 256 1 OMP=
    Process                     = SIGMA_SM_GG_TTXGG_CUDA [nvcc 11.4.120 (gcc 10.3.0)] [inlineHel=0]
    FP precision                = DOUBLE (NaN/abnormal=0, zero=0)
    EvtsPerSec[MatrixElems] (3) = ( 2.983544e+05                 )  sec^-1
    EvtsPerSec[MECalcOnly] (3a) = ( 2.984948e+05                 )  sec^-1
    MeanMatrixElemValue         = ( 4.063123e+00 +- 2.368970e+00 )  GeV^-4
    TOTAL       :     0.632021 sec
           240,479,016      cycles:u                  #    0.271 GHz
           374,344,370      instructions:u            #    1.56  insn per cycle
           0.949804465 seconds time elapsed
    ==PROF== Profiling "sigmaKin": launch__registers_per_thread 255
    ==PROF== Profiling "sigmaKin": sm__sass_average_branch_targets_threads_uniform.pct 100%
    .........................................................................
    runExe /afs/cern.ch/user/a/avalassi/GPU2020/madgraph4gpuX/epochX/cudacpp/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg/build.none_d_inl0/gcheck.exe -p 2048 256 1 OMP=
    Process                     = SIGMA_SM_GG_TTXGG_CUDA [nvcc 11.4.120 (gcc 10.3.0)] [inlineHel=0]
    FP precision                = DOUBLE (NaN/abnormal=0, zero=0)
    EvtsPerSec[MatrixElems] (3) = ( 3.064411e+05                 )  sec^-1
    EvtsPerSec[MECalcOnly] (3a) = ( 3.065075e+05                 )  sec^-1
    MeanMatrixElemValue         = ( 6.665112e+00 +- 5.002651e+00 )  GeV^-4
    TOTAL       :     4.010649 sec
         3,455,121,604      cycles:u                  #    0.807 GHz
         7,063,878,636      instructions:u            #    2.04  insn per cycle
           4.338785554 seconds time elapsed
    =========================================================================
    
    On itscrd70.cern.ch [CPU: Intel(R) Xeon(R) Silver 4216 CPU] [GPU: 1x Tesla V100S-PCIE-32GB]:
    =========================================================================
    runExe /afs/cern.ch/user/a/avalassi/GPU2020/madgraph4gpuX/epochX/cudacpp/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg/build.none_d_inl0/gcheck.exe -p 64 256 1 OMP=
    Process                     = SIGMA_SM_GG_TTXGG_CUDA [nvcc 11.4.120 (gcc 9.2.0)] [inlineHel=0]
    FP precision                = DOUBLE (NaN/abnormal=0, zero=0)
    EvtsPerSec[MatrixElems] (3) = ( 2.957504e+05                 )  sec^-1
    EvtsPerSec[MECalcOnly] (3a) = ( 2.958969e+05                 )  sec^-1
    MeanMatrixElemValue         = ( 4.063123e+00 +- 2.368970e+00 )  GeV^-4
    TOTAL       :     0.641580 sec
           245,717,337      cycles:u                  #    0.275 GHz
           375,031,926      instructions:u            #    1.53  insn per cycle
           0.960705369 seconds time elapsed
    ==PROF== Profiling "sigmaKin": launch__registers_per_thread 255
    ==PROF== Profiling "sigmaKin": sm__sass_average_branch_targets_threads_uniform.pct 100%
    .........................................................................
    runExe /afs/cern.ch/user/a/avalassi/GPU2020/madgraph4gpuX/epochX/cudacpp/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg/build.none_d_inl0/gcheck.exe -p 2048 256 1 OMP=
    Process                     = SIGMA_SM_GG_TTXGG_CUDA [nvcc 11.4.120 (gcc 9.2.0)] [inlineHel=0]
    FP precision                = DOUBLE (NaN/abnormal=0, zero=0)
    EvtsPerSec[MatrixElems] (3) = ( 3.069776e+05                 )  sec^-1
    EvtsPerSec[MECalcOnly] (3a) = ( 3.070445e+05                 )  sec^-1
    MeanMatrixElemValue         = ( 6.665112e+00 +- 5.002651e+00 )  GeV^-4
    TOTAL       :     4.005185 sec
         3,353,639,585      cycles:u                  #    0.785 GHz
         6,992,125,765      instructions:u            #    2.08  insn per cycle
           4.334594570 seconds time elapsed
    =========================================================================
    
    On itscrd70.cern.ch [CPU: Intel(R) Xeon(R) Silver 4216 CPU] [GPU: 1x Tesla V100S-PCIE-32GB]:
    =========================================================================
    runExe /afs/cern.ch/user/a/avalassi/GPU2020/madgraph4gpuX/epochX/cudacpp/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg/build.none_d_inl0/gcheck.exe -p 64 256 1 OMP=
    Process                     = SIGMA_SM_GG_TTXGG_CUDA [nvcc 11.0.221 (gcc 9.2.0)] [inlineHel=0]
    FP precision                = DOUBLE (NaN/abnormal=0, zero=0)
    EvtsPerSec[MatrixElems] (3) = ( 3.688636e+05                 )  sec^-1
    EvtsPerSec[MECalcOnly] (3a) = ( 3.691051e+05                 )  sec^-1
    MeanMatrixElemValue         = ( 4.063123e+00 +- 2.368970e+00 )  GeV^-4
    TOTAL       :     0.605746 sec
           204,463,136      cycles:u                  #    0.240 GHz
           312,854,636      instructions:u            #    1.53  insn per cycle
           0.914451076 seconds time elapsed
    ==PROF== Profiling "sigmaKin": launch__registers_per_thread 255
    ==PROF== Profiling "sigmaKin": sm__sass_average_branch_targets_threads_uniform.pct 100%
    .........................................................................
    runExe /afs/cern.ch/user/a/avalassi/GPU2020/madgraph4gpuX/epochX/cudacpp/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg/build.none_d_inl0/gcheck.exe -p 2048 256 1 OMP=
    Process                     = SIGMA_SM_GG_TTXGG_CUDA [nvcc 11.0.221 (gcc 9.2.0)] [inlineHel=0]
    FP precision                = DOUBLE (NaN/abnormal=0, zero=0)
    EvtsPerSec[MatrixElems] (3) = ( 3.922477e+05                 )  sec^-1
    EvtsPerSec[MECalcOnly] (3a) = ( 3.923566e+05                 )  sec^-1
    MeanMatrixElemValue         = ( 6.665112e+00 +- 5.002651e+00 )  GeV^-4
    TOTAL       :     3.274578 sec
         2,597,597,796      cycles:u                  #    0.735 GHz
         5,608,073,361      instructions:u            #    2.16  insn per cycle
           3.597404062 seconds time elapsed
    =========================================================================
```