@valassi 2021-10-31
The CI tests are still failing but I believe this is a problem with the CI - issue #289. The tests fail in epoch2, while I have not changed epoch2 (or the common test directory), the problem must have been pre-existing. I will merge anyway.