@valassi 2021-10-31
I think the issue in #277 is now understood as much as possible (and as much as it is worth it).

I am about to merge PR #281 with several changes in the code (mainly, the addition of static to cure #283, plus a few fixes to allow older compilers to be used, plus a few performance studies).

In summary, expanding on the post above https://github.com/madgraph5/madgraph4gpu/issues/277#issuecomment-952819519. There are mainly THREE effects:

1) A small effect from using "-p 64 256 1" instead of "-p 2048 256 12".
- was 4.4E5 (64/256/1) to 5.1E5 (2048/256/1) with cuda11.0 and gcc9.2

I will now systematically test also "-p 2048 256 1" (one iteration is enough) for CUDA.
This is also added to the throughputX script, so it will make it to the logs.
See commit fa48545a312d854bbacb8163d7e340e4c746bf1c in PR #281

2) There is a clear performance regression in moving from cuda 11.0/11.1 to cuda 11.2/11.3/11.4/11.5.
With the same code (latest epoch2/ggttgg, equivalent to initial epochx/ggttgg):
- was 4.4E5 (64/256/1) to 5.1E5 (2048/256/1) with cuda11.0 and gcc9.2
- is now 3.5E5 (64/256/1) to 3.8E5 (2048/256/1) with cuda11.4 and gcc10.3

This is analysed in #282. The issue has been reported to NVidia and is being discussed with them.
For the moment, I find it better to stick to an older version of nvcc which provides higher throughput.
The latest nvcc which does not have the issue is cuda 11.1, so I moved to cuda 11.1.
Unlike 11.0, which does not support gcc10, cuda11.1 allows using gcc10.
I have therefore moved to cuda 11.1/gcc10.2 as the new baseline (see results below).
I prefer this to gcc10.3 because I get a few issues on gcc10.3 in a few combinations (maybe a bug in the cvmfs installation).
PR #281 includes several fixes to allow the code to build with these older compilers (e.g. #269, #291).

Note en passant that I wanted to move to c++17 in cuda as I do in gcc.
This however is not possible with cuda 11.1. I have been able to fix #290 but not #292.
So I stick to c++14 for cuda 11.1. For consistency, for the moment I use c++14 also in cuda 11.4.

3) In addition, there is a performance regression from the code itself, from golden tag epochX2 to golden tag epochX4.
(Note that both tags need some fixes to allow building on cuda 110 and gcc 92).

Epoch2
- was 4.4E5 (64/256/1) to 5.1E5 (2048/256/1) with cuda11.0 and gcc9.2
- is now 3.5E5 (64/256/1) to 3.8E5 (2048/256/1) with cuda11.4 and gcc10.3

EpochX before #281
- was 3.7E5 (64/256/1) to 3.9E5 (2048/256/1) with cuda11.0 and gcc9.2
- was 3.0E5 (64/256/1) to 3.1E5 (2048/256/1) with cuda11.4 and gcc10.3

The main issue is a missing static for the color matrix constexpr array. This is now fixed in #281 (see #283), recovering around 30% throughput. There are also a few other smaller issues, for instance the order of TMP lines (#285), which however cannot be fixed in code generation and is probably not worth fixing as it is only 2%. There is probably another regression at the level of 1-2%, but I could not identify it an dit is probably a fluctuation, I will not try to fix it). The final situation is therefore the following, from these three files:
https://github.com/madgraph5/madgraph4gpu/blob/d949fc0c9a11bcb578495acece7584a706c28e8b/epochX/cudacpp/tput/logs_ggttgg_auto/log_ggttgg_auto_d_inl0.txt
https://github.com/madgraph5/madgraph4gpu/blob/d6b0e190033f42e127d630056f6dd1dc22c4ee25/epochX/cudacpp/tput/logs_ggttgg_auto/log_ggttgg_auto_d_inl0.txt
https://github.com/madgraph5/madgraph4gpu/blob/99ec4099ac09836d898edca3ba2478b693ac6ac8/epochX/cudacpp/tput/logs_ggttgg_auto/log_ggttgg_auto_d_inl0.txt

EpochX after #281
- is 4.4E5 (64/256/1) to 4.9E5 (2048/256/1) with cuda11.1/c++14 and gcc10.2/c++17
- is 3.5E5 (64/256/1) to 3.8E5 (2048/256/1) with cuda11.4/c++14 and gcc10.2/c++17
- is 3.5E5 (64/256/1) to 3.8E5 (2048/256/1) with cuda11.4/c++17 and gcc10.2/c++17