@valassi 2021-10-31
Fixed what I could fix, documented. I am self merging